Update main
Fixed the issue regarding the backspace removing entire lines of text. Turns out it had something to do with the clock tick. Also had to add in a conditional to update the screen when backspace was pressed so it would reflect the characters deleted rather than leave the old letters up.